+++ transforms/RTMP/H264Packetizer.cpp
@@ -130,7 +130,7 @@ namespace videocore { namespace rtmp {
             count++;
             
             prev_time = now;
-            outMeta.setData(ts, outBuffer.size(), FLV_TAG_TYPE_VIDEO, kVideoChannelStreamId);
+            outMeta.setData(ts, static_cast<int>(outBuffer.size()), FLV_TAG_TYPE_VIDEO, kVideoChannelStreamId);
             
             output->pushBuffer(&outBuffer[0], outBuffer.size(), outMeta);
         }
